@@ -59,7 +59,7 @@ dependencies:
   scroll_to_index: ^3.0.1
   carousel_slider: ^4.1.1
   scrollable_positioned_list: ^0.3.5
-  webview_flutter: ^3.0.4
+  webview_flutter: ^4.2.0
   photo_view: ^0.14.0
 dev_dependencies:
   flutter_launcher_icons: ^0.10.0
